+++ packages/lumx-core/style-dictionary/config/gen-scss-variables.js
@@ -1,6 +1,6 @@
 const StyleDictionary = require('style-dictionary');
 const path = require('path');
-const transformGroup = require('./utils/_css-transform-group');
+const transformGroup = require('./_transform-group');
 
 /**
  * SCSS generator